KnifeThrowAbility and AttackEmpower start

+++ Assets/_Scripts/Interfaces/IAgent.cs
@@ -2,7 +2,7 @@ using UnityEngine.Events;
 
 public interface IAgent
 {
-    int Health  { get; }
+    float Health  { get; }
     bool IsDead { get; }
     UnityEvent OnDeath { get; set; }
     UnityEvent OnGetHit { get; set; }
